Fixes Gatsby build error when running in a non-browser environment.
Why: Running gatsby build locally or in prod fails because the sessionStorage variable is not available. This happens because the build command is run in a shell where the browser APIs are not available and hence causing and undefined sessionStorage variable.

Fix: Check if 'window' is defined before accessing sessionStorage